fix: moved types to a types folder
