--- client/src/Models/RentalModel.tsx
@@ -1,7 +1,0 @@
-export interface Rental {
-    id: number;
-    item_id: number;
-    renter_id: number;
-    start_date: string;
-    end_date: string;
-}
